// api/src/modules/auth/auth.schemas.ts
import { z } from 'zod';

export const registerSchema = z.object({
  name: z.string().min(2).max(80).optional(),
  email: z.string().email().toLowerCase(),
  password: z.string().min(8).max(100),
});

export const loginSchema = z.object({
  email: z.string().email().toLowerCase(),
  password: z.string().min(1),
});

export type RegisterInput = z.infer<typeof registerSchema>;
export type LoginInput = z.infer<typeof loginSchema>;